Fix bugs on route end: a button appears if the user does not have an internet connection when he clicks on the stop button Fix average speed update when changing orientation on pause Stop RouteService even if the user does not have an internet connection on route end

+++ mobile/src/main/java/brice/explorun/fragments/MapFragment.java
@@ -606,11 +606,11 @@ public class MapFragment extends PlacesObserverFragment implements OnMapReadyCal
 			List<LatLng> locations = new ArrayList<>();
 			ArrayList<Step> instructions = new ArrayList<>();
 
-			float distance = 0;
+			double distance = 0;
 
 			for (Leg leg: route.getLegList())
 			{
-				distance += Float.parseFloat(leg.getDistance().getValue());
+				distance += Double.parseDouble(leg.getDistance().getValue());
 				Coordination coordination = leg.getStartLocation();
 				locations.add(new LatLng(coordination.getLatitude(), coordination.getLongitude()));
 				for(Step step : leg.getStepList()){
